refactor(test) : Test클래스 위치는 본체 패키지 위치와 동일하게
